requirements.txt
absl-py==0.9.0
appdirs==1.4.3
appnope==0.1.0
astroid==2.3.3
astropy==4.0.1.post1
astunparse==1.6.3
attrs==19.3.0
autopep8==1.5
backcall==0.1.0
black==19.10b0
bleach==3.1.5
Brotli==1.0.7
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
chart-studio==1.1.0
click==7.1.1
cmdstanpy==0.4.0
colorama==0.4.3
colorlover==0.3.0
configparser==5.0.0
confuse==1.1.0
cufflinks==0.17.3
cycler==0.10.0
Cython==0.29.17
dash==1.11.0
dash-core-components==1.9.1
dash-html-components==1.0.3
dash-renderer==1.4.0
dash-table==4.6.2
decorator==4.4.2
defusedxml==0.6.0
docker-pycreds==0.4.0
entrypoints==0.3
fastapi==0.54.2
Flask==1.1.1
Flask-Compress==1.5.0
Flask-SQLAlchemy==2.4.1
future==0.18.2
gast==0.3.3
gitdb==4.0.4
GitPython==3.1.1
google-auth==1.13.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
gql==0.2.0
graphql-core==1.1
grpcio==1.28.1
h5py==2.10.0
htmlmin==0.1.12
idna==2.9
ImageHash==4.1.0
importlib-metadata==1.6.0
ipykernel==5.2.1
ipython==7.14.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
itsdangerous==1.1.0
jedi==0.17.0
Jinja2==2.11.2
joblib==0.14.1
json5==0.8.5
jsonschema==3.2.0
jupyter-client==6.1.3
jupyter-core==4.6.3
jupyterlab==1.2.5
jupyterlab-server==1.0.6
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keras-tuner==1.0.1
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
llvmlite==0.32.1
Markdown==3.2.1
MarkupSafe==1.1.1
matplotlib==3.2.1
mccabe==0.6.1
missingno==0.4.2
mistune==0.8.4
mlxtend==0.17.2
mysql-connector-python==8.0.20
nbconvert==5.6.1
nbformat==5.0.6
networkx==2.4
notebook==6.0.3
numba==0.49.1
numpy==1.18.4
nvidia-ml-py3==7.352.0
oauthlib==3.1.0
opt-einsum==3.2.0
packaging==20.3
pandas==1.0.3
pandas-profiling==2.8.0
pandocfilters==1.4.2
parso==0.7.0
pathspec==0.7.0
pathtools==0.1.2
patsy==0.5.1
pexpect==4.8.0
phik==0.9.12
pi==0.1.2
pickleshare==0.7.5
Pillow==7.1.2
plotly==4.6.0
prometheus-client==0.7.1
promise==2.3
prompt-toolkit==3.0.5
protobuf==3.11.3
psutil==5.7.0
psycopg2-binary==2.8.4
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pydantic==1.5.1
Pygments==2.6.1
pylint==2.4.4
PyMySQL==0.9.3
pymysql2==1.2.6
pyparsing==2.4.7
pyrsistent==0.16.0
pystan==2.19.1.1
python-dateutil==2.8.1
python-dotenv==0.13.0
pytz==2020.1
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==19.0.1
qgrid==1.3.1
regex==2020.2.20
requests==2.23.0
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.0
scikit-learn==0.22.2.post1
scipy==1.4.1
seaborn==0.10.0
Send2Trash==1.5.0
sentry-sdk==0.14.3
shortuuid==1.0.1
six==1.14.0
smmap==3.0.2
spotipy==2.12.0
SQLAlchemy==1.3.16
starlette==0.13.2
statsmodels==0.11.1
subprocess32==3.5.4
tabulate==0.8.7
tangled-up-in-unicode==0.0.6
tensorboard==2.2.0
tensorboard-plugin-wit==1.6.0.post3
tensorflow==2.2.0rc3
tensorflow-estimator==2.2.0rc0
termcolor==1.1.0
terminado==0.8.3
terminaltables==3.1.0
testpath==0.4.4
toml==0.10.0
tornado==6.0.4
tqdm==4.45.0
traitlets==4.3.3
typed-ast==1.4.1
urllib3==1.25.8
virtualenv==16.7.9
visions==0.4.4
wandb==0.8.31
watchdog==0.10.2
wcwidth==0.1.9
webencodings==0.5.1
Werkzeug==1.0.0
widgetsnbextension==3.5.1
wrapt==1.11.2
zipp==3.1.0
